fix: Registry dependency
Changed registry from a hard dependency to a dev dep/peer dep. Peer dep should be sufficient, assuming that the version will be pulled in by the platform/workspace, but since our apps can build standalone and Yarn does not pull in peer deps by default, it has been added as a dev dep as well. Versions will need to be kept aligned with any other Registry dependencies in a build to ensure that multiple Registries aren't being registered to/read from.

ERM-2073